functions/bronze_ingest/requirements.txt
google-cloud-storage>=3.0.0
google-cloud-bigquery>=3.0.0
google-cloud-pubsub>=2.0.0
